remove installation steps for wxWidgets & Perl
wxWidgets is now downloaded & built automatically from CMake.

Perl is not required by CMake - it searches for the interpreter
but the build succeeds even if it does not find it. Same goes
for Python, but as we use Python in other places I think it's
still best to keep it.